pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mergescope"
version = "0.1.0"
description = "Audit merged PRs against Jira tickets using a LangGraph agent and MCP tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Martin Caminoa" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "langgraph>=0.4,<2.0",
    "langchain-mcp-adapters>=0.1,<2.0",
    "langchain-anthropic>=0.3,<2.0",
    "langchain-aws>=0.2,<2.0",
    "langchain-openai>=0.3,<2.0",
    "pyyaml>=6.0,<7.0",
    "rich>=13.0,<15.0",
]

[project.scripts]
mergescope = "mergescope.cli:main"

[project.urls]
Homepage = "https://github.com/martin5211/MergeScope"
Repository = "https://github.com/martin5211/MergeScope"
Issues = "https://github.com/martin5211/MergeScope/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mergescope"]

[tool.pytest.ini_options]
testpaths = ["tests"]
